Move MainWindowWithDetachableDockWidgets to separate file
